+++ index.ts
@@ -5,28 +5,27 @@ import {LogType} from '@src/logger/logger-export';
 
 globalThis.hb = new TwitchBot()
 
-{
-  ;(async () => {
-    hb.log(LogType.INFO, 'Initializing...')
-    await hb.init()
+    {
+      ;(async () => {
+        hb.log(LogType.INFO, 'Initializing...')
+        await hb.init()
 
-    if (hb.config.isDev()) {
-      await setupDev()
+        if (hb.config.isDev()) {
+          await setupDev()
+        }
+
+        hb.startJobs()
+        hb.initModules()
+      })()
     }
 
-    hb.startJobs()
-    hb.initModules()
-  })()
-}
+    process.on('uncaughtException', async error => {
+      await hb.db.error.save({
+        message: error.message,
+        stack_trace: error.stack,
+        timestamp: Date.now()
+      })
+      console.error(error)
 
-process.on('uncaughtException', async error => {
-  await hb.db.error.save({
-    message: error.message,
-    stack_trace: error.stack,
-    timestamp: Date.now()
-  })
-  console.error(error)
-
-  process.exit(1)
-})
-
+      process.exit(1)
+    })