+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.1](https://github.com/helltf/helltfbot-v2/releases/tag/v1.4.1) - 2022-11-11
+
+### Added
+
+- notification on suggestion update
+
+### Changed
+
+- github pipeline updates uses webhook
+- updated debug logs
+
 ## [1.4.0](https://github.com/helltf/helltfbot-v2/releases/tag/v1.4.0) - 2022-10-13
 
 ### Added